Adds base functionality
