+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/clientSide.tsx
@@ -14,10 +14,12 @@ export const FormWrapper = ({
   formRecord,
   currentForm,
   allowGrouping,
+  hCaptchaSiteKey,
 }: {
   formRecord: TypeOmit<FormRecord, "name" | "deliveryOption">;
   currentForm: JSX.Element[];
   allowGrouping?: boolean | undefined;
+  hCaptchaSiteKey: string;
 }) => {
   // TODO cast language as "en" | "fr" in TS below
   const {
@@ -61,6 +63,7 @@ export const FormWrapper = ({
         );
       }}
       allowGrouping={allowGrouping}
+      hCaptchaSiteKey={hCaptchaSiteKey}
     >
       {currentForm}
     </Form>
