+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/page.tsx
@@ -13,6 +13,7 @@ import { allowGrouping } from "@formBuilder/components/shared/right-panel/treevi
 import { serverTranslation } from "@i18n";
 import { ClosingNotice } from "@clientComponents/forms/ClosingNotice/ClosingNotice";
 import { FormDelayProvider } from "@lib/hooks/useFormDelayContext";
+import { getHCaptchaSettings } from "@clientComponents/globals/Captcha/helpers";
 
 export async function generateMetadata(props0: {
   params: Promise<{ locale: string; props: string[] }>;
@@ -65,6 +66,8 @@ export default async function Page(props0: {
   const formTitle = formRecord.form[getLocalizedProperty("title", language)] as string;
   const isAllowGrouping = allowGrouping();
 
+  const { hCaptchaSiteKey } = await getHCaptchaSettings();
+
   let isPastClosingDate = false;
 
   if (formRecord.closingDate) {
@@ -101,6 +104,7 @@ export default async function Page(props0: {
               formRecord={formRecord}
               currentForm={currentForm}
               allowGrouping={isAllowGrouping}
+              hCaptchaSiteKey={hCaptchaSiteKey}
             />
           </FormDelayProvider>
         </GCFormsProvider>
