+++ lib/cspScripts.ts
@@ -27,21 +27,38 @@ if (window.location.host === "forms-formulaires.alpha.canada.ca") {
 export const generateCSP = (): { csp: string; nonce: string } => {
   const nonce = Buffer.from(crypto.randomUUID()).toString("base64");
 
+  //   const cspHeader = `
+  //     default-src 'self';
+  //     script-src 'self' 'nonce-${nonce}' 'strict-dynamic';
+  //     style-src 'self' 'nonce-${nonce}';
+  //     img-src 'self' blob: data:;
+  //     font-src 'self';
+  //     object-src 'none';
+  //     base-uri 'self';
+  //     form-action 'self';
+  //     frame-src www.googletagmanager.com;
+  //     frame-ancestors 'none';
+  //     connect-src 'self' www.googletagmanager.com www.google-analytics.com ws1.postescanada-canadapost.ca;
+  //     block-all-mixed-content;
+  //     upgrade-insecure-requests;
+  // `;
+
+  // TODO self-host hCAPTCHA script and revert to original (above)
   const cspHeader = `
-    default-src 'self';
-    script-src 'self' 'nonce-${nonce}' 'strict-dynamic';
-    style-src 'self' 'nonce-${nonce}';
-    img-src 'self' blob: data:;
-    font-src 'self';
-    object-src 'none';
-    base-uri 'self';
-    form-action 'self';
-    frame-src www.googletagmanager.com;
-    frame-ancestors 'none';
-    connect-src 'self' www.googletagmanager.com www.google-analytics.com ws1.postescanada-canadapost.ca;
-    block-all-mixed-content;
-    upgrade-insecure-requests;
-`;
+      default-src 'self';
+      script-src 'self' 'nonce-${nonce}' 'strict-dynamic';
+      style-src 'self' 'nonce-${nonce}';
+      img-src 'self' blob: data:;
+      font-src 'self';
+      object-src 'none';
+      base-uri 'self';
+      form-action 'self';
+      frame-src www.googletagmanager.com https://hcaptcha.com https://*.hcaptcha.com;
+      frame-ancestors 'none';
+      connect-src 'self' www.googletagmanager.com www.google-analytics.com ws1.postescanada-canadapost.ca https://hcaptcha.com https://*.hcaptcha.com;
+      block-all-mixed-content;
+      upgrade-insecure-requests;
+    `;
 
   // replace newline characters with spaces
   return { csp: cspHeader.replace(/\s{2,}/g, " ").trim(), nonce };
